[test] JWT 인증 로직 구현 전 Unit test code 틀 완성
Motivation:

 - #95 를 참고하세요.
