+++ src/test/java/com/kdt_y_be_toy_project2/global/factory/ItineraryTestFactory.java
@@ -1,14 +1,20 @@
 package com.kdt_y_be_toy_project2.global.factory;
 
 import com.kdt_y_be_toy_project2.domain.itinerary.domain.*;
+import com.kdt_y_be_toy_project2.domain.itinerary.dto.ItineraryRequest;
+import com.kdt_y_be_toy_project2.domain.itinerary.dto.request.AccommodationInfoRequest;
+import com.kdt_y_be_toy_project2.domain.itinerary.dto.request.MoveInfoRequest;
+import com.kdt_y_be_toy_project2.domain.itinerary.dto.request.StayInfoRequest;
+import com.kdt_y_be_toy_project2.domain.model.DateTimeScheduleInfo;
 import com.kdt_y_be_toy_project2.domain.model.PlaceInfo;
 import com.kdt_y_be_toy_project2.domain.trip.domain.Trip;
+import com.kdt_y_be_toy_project2.global.util.LocalDateTimeUtil;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-public class ItineraryFactory {
+public class ItineraryTestFactory {
 
 
     public static Itinerary createTestItinerary() {
@@ -48,6 +54,25 @@ public class ItineraryFactory {
                         .staySchedule(DateTimeScheduleInfoTestFactory.createRandom())
                         .build())
                 .trip(trip)
+                .build();
+    }
+
+    public static ItineraryRequest buildItineraryRequest(DateTimeScheduleInfo Schedule) {
+        return ItineraryRequest.builder()
+                .accommodationInfoRequest(AccommodationInfoRequest.builder()
+                        .startDateTime(LocalDateTimeUtil.toString(Schedule.getStartDateTime()))
+                        .endDateTime(LocalDateTimeUtil.toString(Schedule.getEndDateTime()))
+                        .accommodationPlaceName("test").build())
+                .stayInfoRequest(StayInfoRequest.builder()
+                        .startDateTime(LocalDateTimeUtil.toString(Schedule.getStartDateTime()))
+                        .endDateTime(LocalDateTimeUtil.toString(Schedule.getEndDateTime()))
+                        .stayPlaceName("test").build())
+                .moveInfoRequest(MoveInfoRequest.builder()
+                        .startDateTime(LocalDateTimeUtil.toString(Schedule.getStartDateTime()))
+                        .endDateTime(LocalDateTimeUtil.toString(Schedule.getEndDateTime()))
+                        .sourcePlaceName("test")
+                        .destPlaceName("test")
+                        .transportationType(TransportationType.WALK.getValue()).build())
                 .build();
     }
